Fix inconsistency in "askai" case sensitivity
This fixes a bug where detection of the "askai" was case insensitive,
but the line that remove "askai" from the LLM prompt was case sensitive.

This meant that if it was spelled "Askai" (commonly capitalized
automatically by mobile keyboards), it would pass the word "Askai" to
the LLM.

There's no simple pre-built Python library for case insensitive replace,
so I rolled my own utility for it.